remove redundant code
